+++ src/components/Emoji/EmojiBadge.jsx
@@ -1,0 +1,5 @@
+import styled from 'styled-components';
+const Container = styled.div``;
+export const EmojiBadge = () => {
+  return <Container>EmojiBadge</Container>;
+};
