+++ src/components/common/Card/PaperCard.jsx
@@ -86,7 +86,7 @@ export const RelationShip = styled.div`
 export function PaperCard({ sender }) {
   if (!sender) return null;
 
-  const { content, createdAt, font, id, profileImageURL, recipientId, relationship, name } = sender;
+  const { content, createdAt, font, profileImageURL, relationship, name } = sender;
 
   const formattedDate = new Date(createdAt).toLocaleDateString();
   return (
